+++ controllers/controllers/workloads/cfapp_controller.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
 
 	korifiv1alpha1 "code.cloudfoundry.org/korifi/controllers/api/v1alpha1"
 	"code.cloudfoundry.org/korifi/controllers/controllers/shared"
@@ -22,11 +23,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/controller-runtime/pkg/source"
-)
-
-const (
-	cfAppFinalizerName = "cfApp.korifi.cloudfoundry.org"
 )
 
 type EnvValueBuilder interface {
@@ -53,73 +49,121 @@ func NewCFAppReconciler(k8sClient client.Client, scheme *runtime.Scheme, log log
 	return k8s.NewPatchingReconciler[korifiv1alpha1.CFApp, *korifiv1alpha1.CFApp](log, k8sClient, &appReconciler)
 }
 
-//+kubebuilder:rbac:groups=korifi.cloudfoundry.org,resources=cfapps,verbs=get;list;watch;create;update;patch;delete
+func (r *CFAppReconciler) SetupWithManager(mgr ctrl.Manager) *builder.Builder {
+	return ctrl.NewControllerManagedBy(mgr).
+		For(&korifiv1alpha1.CFApp{}).
+		Watches(
+			&korifiv1alpha1.CFBuild{},
+			handler.EnqueueRequestsFromMapFunc(buildToApp),
+		).
+		Watches(
+			&korifiv1alpha1.CFServiceBinding{},
+			handler.EnqueueRequestsFromMapFunc(serviceBindingToApp),
+		)
+}
+
+func buildToApp(ctx context.Context, o client.Object) []reconcile.Request {
+	cfBuild, ok := o.(*korifiv1alpha1.CFBuild)
+	if !ok {
+		return nil
+	}
+
+	return []reconcile.Request{
+		{
+			NamespacedName: types.NamespacedName{
+				Name:      cfBuild.Spec.AppRef.Name,
+				Namespace: o.GetNamespace(),
+			},
+		},
+	}
+}
+
+func serviceBindingToApp(ctx context.Context, o client.Object) []reconcile.Request {
+	serviceBinding, ok := o.(*korifiv1alpha1.CFServiceBinding)
+	if !ok {
+		return nil
+	}
+
+	return []reconcile.Request{
+		{
+			NamespacedName: types.NamespacedName{
+				Name:      serviceBinding.Spec.AppRef.Name,
+				Namespace: o.GetNamespace(),
+			},
+		},
+	}
+}
+
+//+kubebuilder:rbac:groups=korifi.cloudfoundry.org,resources=cfapps,verbs=get;list;watch;create;update;patch;delete;deletecollection
 //+kubebuilder:rbac:groups=korifi.cloudfoundry.org,resources=cfapps/status,verbs=get;update;patch
 //+kubebuilder:rbac:groups=korifi.cloudfoundry.org,resources=cfapps/finalizers,verbs=update
+
 //+kubebuilder:rbac:groups="",resources=secrets,verbs=get;list;watch;patch
 
 func (r *CFAppReconciler) ReconcileResource(ctx context.Context, cfApp *korifiv1alpha1.CFApp) (ctrl.Result, error) {
-	log := r.log.WithValues("namespace", cfApp.Namespace, "name", cfApp.Name)
+	log := shared.ObjectLogger(r.log, cfApp)
+	ctx = logr.NewContext(ctx, log)
+
+	cfApp.Status.ObservedGeneration = cfApp.Generation
+	log.V(1).Info("set observed generation", "generation", cfApp.Status.ObservedGeneration)
 
 	if !cfApp.GetDeletionTimestamp().IsZero() {
-		err := r.finalizeCFApp(ctx, log, cfApp)
-		return ctrl.Result{}, err
+		return r.finalizeCFApp(ctx, cfApp)
 	}
 
-	err := k8s.AddFinalizer(ctx, log, r.k8sClient, cfApp, cfAppFinalizerName)
-	if err != nil {
-		log.Info("error adding finalizer", "reason", err)
-		return ctrl.Result{}, err
+	if cfApp.Annotations[korifiv1alpha1.CFAppLastStopRevisionKey] == "" {
+		if cfApp.Annotations == nil {
+			cfApp.Annotations = map[string]string{}
+		}
+		cfApp.Annotations[korifiv1alpha1.CFAppLastStopRevisionKey] = cfApp.Annotations[korifiv1alpha1.CFAppRevisionKey]
 	}
 
 	secretName := cfApp.Name + "-vcap-application"
-	err = r.reconcileVCAPSecret(ctx, log, cfApp, secretName, r.vcapApplicationEnvBuilder)
+	err := r.reconcileVCAPSecret(ctx, cfApp, secretName, r.vcapApplicationEnvBuilder)
 	if err != nil {
-		log.Info("unable to create CFApp VCAP Application secret", "reason", err)
 		return ctrl.Result{}, err
 	}
 	cfApp.Status.VCAPApplicationSecretName = secretName
 
 	secretName = cfApp.Name + "-vcap-services"
-	err = r.reconcileVCAPSecret(ctx, log, cfApp, secretName, r.vcapServicesEnvBuilder)
+	err = r.reconcileVCAPSecret(ctx, cfApp, secretName, r.vcapServicesEnvBuilder)
 	if err != nil {
-		log.Info("unable to create CFApp VCAP Services secret", "reason", err)
 		return ctrl.Result{}, err
 	}
+
 	cfApp.Status.VCAPServicesSecretName = secretName
 
-	if cfApp.Status.Conditions == nil {
-		cfApp.Status.Conditions = make([]metav1.Condition, 0)
-	}
-
-	meta.SetStatusCondition(&cfApp.Status.Conditions, metav1.Condition{
-		Type:   StatusConditionReady,
-		Status: metav1.ConditionFalse,
-		Reason: "DropletNotAssigned",
-	})
-
 	if cfApp.Spec.CurrentDropletRef.Name == "" {
+		meta.SetStatusCondition(&cfApp.Status.Conditions, metav1.Condition{
+			Type:               shared.StatusConditionReady,
+			Status:             metav1.ConditionFalse,
+			Reason:             "DropletNotAssigned",
+			ObservedGeneration: cfApp.Generation,
+		})
+
 		return ctrl.Result{}, nil
 	}
 
-	meta.SetStatusCondition(&cfApp.Status.Conditions, metav1.Condition{
-		Type:   StatusConditionReady,
-		Status: metav1.ConditionFalse,
-		Reason: "CannotResolveCurrentDropletRef",
-	})
-
-	droplet, err := r.getDroplet(ctx, log, cfApp)
+	droplet, err := r.getDroplet(ctx, cfApp)
 	if err != nil {
+		meta.SetStatusCondition(&cfApp.Status.Conditions, metav1.Condition{
+			Type:               shared.StatusConditionReady,
+			Status:             metav1.ConditionFalse,
+			Reason:             "CannotResolveCurrentDropletRef",
+			ObservedGeneration: cfApp.Generation,
+		})
+
 		return ctrl.Result{}, err
 	}
 
 	meta.SetStatusCondition(&cfApp.Status.Conditions, metav1.Condition{
-		Type:   StatusConditionReady,
-		Status: metav1.ConditionTrue,
-		Reason: "DropletAssigned",
+		Type:               shared.StatusConditionReady,
+		Status:             metav1.ConditionTrue,
+		Reason:             "DropletAssigned",
+		ObservedGeneration: cfApp.Generation,
 	})
 
-	err = r.startApp(ctx, log, cfApp, droplet)
+	err = r.startApp(ctx, cfApp, droplet)
 	if err != nil {
 		return ctrl.Result{}, err
 	}
@@ -127,8 +171,8 @@ func (r *CFAppReconciler) ReconcileResource(ctx context.Context, cfApp *korifiv1
 	return ctrl.Result{}, nil
 }
 
-func (r *CFAppReconciler) getDroplet(ctx context.Context, log logr.Logger, cfApp *korifiv1alpha1.CFApp) (*korifiv1alpha1.BuildDropletStatus, error) {
-	log = log.WithName("getDroplet").WithValues("dropletName", cfApp.Spec.CurrentDropletRef.Name)
+func (r *CFAppReconciler) getDroplet(ctx context.Context, cfApp *korifiv1alpha1.CFApp) (*korifiv1alpha1.BuildDropletStatus, error) {
+	log := logr.FromContextOrDiscard(ctx).WithName("getDroplet").WithValues("dropletName", cfApp.Spec.CurrentDropletRef.Name)
 
 	var cfBuild korifiv1alpha1.CFBuild
 	err := r.k8sClient.Get(ctx, types.NamespacedName{Name: cfApp.Spec.CurrentDropletRef.Name, Namespace: cfApp.Namespace}, &cfBuild)
@@ -146,15 +190,16 @@ func (r *CFAppReconciler) getDroplet(ctx context.Context, log logr.Logger, cfApp
 	return cfBuild.Status.Droplet, nil
 }
 
-func (r *CFAppReconciler) startApp(ctx context.Context, log logr.Logger, cfApp *korifiv1alpha1.CFApp, droplet *korifiv1alpha1.BuildDropletStatus) error {
-	log = log.WithName("startApp")
+func (r *CFAppReconciler) startApp(ctx context.Context, cfApp *korifiv1alpha1.CFApp, droplet *korifiv1alpha1.BuildDropletStatus) error {
+	log := logr.FromContextOrDiscard(ctx).WithName("startApp")
 
 	for _, dropletProcess := range addWebIfMissing(droplet.ProcessTypes) {
 		loopLog := log.WithValues("processType", dropletProcess.Type)
+		ctx = logr.NewContext(ctx, loopLog)
 
-		existingProcess, err := r.fetchProcessByType(ctx, log, cfApp.Name, cfApp.Namespace, dropletProcess.Type)
+		existingProcess, err := r.fetchProcessByType(ctx, cfApp.Name, cfApp.Namespace, dropletProcess.Type)
 		if err != nil {
-			loopLog.Info("error when fetching  cfprocess by type", "reason", err)
+			loopLog.Info("error when fetching CFProcess by type", "reason", err)
 			return err
 		}
 
@@ -165,7 +210,7 @@ func (r *CFAppReconciler) startApp(ctx context.Context, log logr.Logger, cfApp *
 				return err
 			}
 		} else {
-			err = r.createCFProcess(ctx, loopLog, dropletProcess, droplet.Ports, cfApp)
+			err = r.createCFProcess(ctx, dropletProcess, droplet.Ports, cfApp)
 			if err != nil {
 				loopLog.Info("error creating CFProcess", "reason", err)
 				return err
@@ -182,6 +227,7 @@ func addWebIfMissing(processTypes []korifiv1alpha1.ProcessType) []korifiv1alpha1
 			return processTypes
 		}
 	}
+
 	return append([]korifiv1alpha1.ProcessType{{Type: korifiv1alpha1.ProcessTypeWeb}}, processTypes...)
 }
 
@@ -191,7 +237,7 @@ func (r *CFAppReconciler) updateCFProcessCommand(ctx context.Context, process *k
 	})
 }
 
-func (r *CFAppReconciler) createCFProcess(ctx context.Context, log logr.Logger, process korifiv1alpha1.ProcessType, ports []int32, cfApp *korifiv1alpha1.CFApp) error {
+func (r *CFAppReconciler) createCFProcess(ctx context.Context, process korifiv1alpha1.ProcessType, ports []int32, cfApp *korifiv1alpha1.CFApp) error {
 	desiredCFProcess := &korifiv1alpha1.CFProcess{
 		ObjectMeta: metav1.ObjectMeta{
 			Namespace: cfApp.Namespace,
@@ -217,7 +263,7 @@ func (r *CFAppReconciler) createCFProcess(ctx context.Context, log logr.Logger, 
 	return r.k8sClient.Create(ctx, desiredCFProcess)
 }
 
-func (r *CFAppReconciler) fetchProcessByType(ctx context.Context, log logr.Logger, appGUID, appNamespace, processType string) (*korifiv1alpha1.CFProcess, error) {
+func (r *CFAppReconciler) fetchProcessByType(ctx context.Context, appGUID, appNamespace, processType string) (*korifiv1alpha1.CFProcess, error) {
 	selector, err := labels.ValidatedSelectorFromSet(map[string]string{
 		korifiv1alpha1.CFAppGUIDLabelKey:     appGUID,
 		korifiv1alpha1.CFProcessTypeLabelKey: processType,
@@ -241,37 +287,41 @@ func (r *CFAppReconciler) fetchProcessByType(ctx context.Context, log logr.Logge
 	return &cfProcessList.Items[0], nil
 }
 
-func (r *CFAppReconciler) finalizeCFApp(ctx context.Context, log logr.Logger, cfApp *korifiv1alpha1.CFApp) error {
-	log = log.WithName("finalize")
+func (r *CFAppReconciler) finalizeCFApp(ctx context.Context, cfApp *korifiv1alpha1.CFApp) (ctrl.Result, error) {
+	log := logr.FromContextOrDiscard(ctx).WithName("finalizeCFApp")
 
-	if !controllerutil.ContainsFinalizer(cfApp, cfAppFinalizerName) {
-		return nil
+	if !controllerutil.ContainsFinalizer(cfApp, korifiv1alpha1.CFAppFinalizerName) {
+		return ctrl.Result{}, nil
 	}
 
-	err := r.finalizeCFAppRoutes(ctx, log, cfApp)
+	err := r.finalizeCFAppRoutes(ctx, cfApp)
 	if err != nil {
-		return err
+		return ctrl.Result{}, err
 	}
 
-	err = r.finalizeCFServiceBindings(ctx, log, cfApp)
+	sbFinalizationResult, err := r.finalizeCFServiceBindings(ctx, cfApp)
 	if err != nil {
-		return err
+		return ctrl.Result{}, err
 	}
 
-	if controllerutil.RemoveFinalizer(cfApp, cfAppFinalizerName) {
+	if (sbFinalizationResult != ctrl.Result{}) {
+		return sbFinalizationResult, nil
+	}
+
+	if controllerutil.RemoveFinalizer(cfApp, korifiv1alpha1.CFAppFinalizerName) {
 		log.V(1).Info("finalizer removed")
 	}
 
-	return nil
+	return ctrl.Result{}, nil
 }
 
-func (r *CFAppReconciler) finalizeCFAppRoutes(ctx context.Context, log logr.Logger, cfApp *korifiv1alpha1.CFApp) error {
-	cfRoutes, err := r.getCFRoutes(ctx, log, cfApp.Name, cfApp.Namespace)
+func (r *CFAppReconciler) finalizeCFAppRoutes(ctx context.Context, cfApp *korifiv1alpha1.CFApp) error {
+	cfRoutes, err := r.getCFRoutes(ctx, cfApp.Name, cfApp.Namespace)
 	if err != nil {
 		return err
 	}
 
-	err = r.updateRouteDestinations(ctx, log, cfApp.Name, cfRoutes)
+	err = r.updateRouteDestinations(ctx, cfApp.Name, cfRoutes)
 	if err != nil {
 		return err
 	}
@@ -279,27 +329,33 @@ func (r *CFAppReconciler) finalizeCFAppRoutes(ctx context.Context, log logr.Logg
 	return nil
 }
 
-func (r *CFAppReconciler) finalizeCFServiceBindings(ctx context.Context, log logr.Logger, cfApp *korifiv1alpha1.CFApp) error {
+func (r *CFAppReconciler) finalizeCFServiceBindings(ctx context.Context, cfApp *korifiv1alpha1.CFApp) (ctrl.Result, error) {
+	log := logr.FromContextOrDiscard(ctx).WithName("finalizeCFServiceBindings")
+
 	sbList := korifiv1alpha1.CFServiceBindingList{}
 	err := r.k8sClient.List(ctx, &sbList, client.InNamespace(cfApp.Namespace), client.MatchingFields{shared.IndexServiceBindingAppGUID: cfApp.Name})
 	if err != nil {
 		log.Info("failed to list app service bindings", "reason", err)
-		return err
+		return ctrl.Result{}, err
+	}
+
+	if len(sbList.Items) == 0 {
+		return ctrl.Result{}, nil
 	}
 
 	for i := range sbList.Items {
 		err = r.k8sClient.Delete(ctx, &sbList.Items[i])
 		if err != nil {
 			log.Info("failed to delete service binding", "serviceBindingName", sbList.Items[i].Name, "reason", err)
-			return err
+			return ctrl.Result{}, err
 		}
 	}
 
-	return nil
+	return ctrl.Result{RequeueAfter: time.Second}, nil
 }
 
-func (r *CFAppReconciler) updateRouteDestinations(ctx context.Context, log logr.Logger, cfAppGUID string, cfRoutes []korifiv1alpha1.CFRoute) error {
-	log = log.WithName("updateRouteDestinations")
+func (r *CFAppReconciler) updateRouteDestinations(ctx context.Context, cfAppGUID string, cfRoutes []korifiv1alpha1.CFRoute) error {
+	log := logr.FromContextOrDiscard(ctx).WithName("updateRouteDestinations")
 
 	for i := range cfRoutes {
 		loopLog := log.WithValues("routeName", cfRoutes[i].Name)
@@ -326,7 +382,9 @@ func (r *CFAppReconciler) updateRouteDestinations(ctx context.Context, log logr.
 	return nil
 }
 
-func (r *CFAppReconciler) getCFRoutes(ctx context.Context, log logr.Logger, cfAppGUID string, cfAppNamespace string) ([]korifiv1alpha1.CFRoute, error) {
+func (r *CFAppReconciler) getCFRoutes(ctx context.Context, cfAppGUID string, cfAppNamespace string) ([]korifiv1alpha1.CFRoute, error) {
+	log := logr.FromContextOrDiscard(ctx).WithName("getCFRoutes")
+
 	var foundRoutes korifiv1alpha1.CFRouteList
 	matchingFields := client.MatchingFields{shared.IndexRouteDestinationAppName: cfAppGUID}
 	err := r.k8sClient.List(context.Background(), &foundRoutes, client.InNamespace(cfAppNamespace), matchingFields)
@@ -338,55 +396,13 @@ func (r *CFAppReconciler) getCFRoutes(ctx context.Context, log logr.Logger, cfAp
 	return foundRoutes.Items, nil
 }
 
-func (r *CFAppReconciler) SetupWithManager(mgr ctrl.Manager) *builder.Builder {
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&korifiv1alpha1.CFApp{}).
-		Watches(&source.Kind{Type: &korifiv1alpha1.CFBuild{}}, handler.EnqueueRequestsFromMapFunc(buildToApp)).
-		Watches(&source.Kind{Type: &korifiv1alpha1.CFServiceBinding{}}, handler.EnqueueRequestsFromMapFunc(serviceBindingToApp))
-}
-
-func buildToApp(o client.Object) []reconcile.Request {
-	cfBuild, ok := o.(*korifiv1alpha1.CFBuild)
-	if !ok {
-		return nil
-	}
-
-	return []reconcile.Request{
-		{
-			NamespacedName: types.NamespacedName{
-				Name:      cfBuild.Spec.AppRef.Name,
-				Namespace: o.GetNamespace(),
-			},
-		},
-	}
-}
-
-func serviceBindingToApp(o client.Object) []reconcile.Request {
-	serviceBinding, ok := o.(*korifiv1alpha1.CFServiceBinding)
-	if !ok {
-		return nil
-	}
-
-	result := []reconcile.Request{
-		{
-			NamespacedName: types.NamespacedName{
-				Name:      serviceBinding.Spec.AppRef.Name,
-				Namespace: o.GetNamespace(),
-			},
-		},
-	}
-
-	return result
-}
-
 func (r *CFAppReconciler) reconcileVCAPSecret(
 	ctx context.Context,
-	log logr.Logger,
 	cfApp *korifiv1alpha1.CFApp,
 	secretName string,
 	envBuilder EnvValueBuilder,
 ) error {
-	log = log.WithName("reconcileVCAPSecret").WithValues("secretName", secretName)
+	log := logr.FromContextOrDiscard(ctx).WithName("reconcileVCAPSecret").WithValues("secretName", secretName)
 
 	secret := &corev1.Secret{
 		ObjectMeta: metav1.ObjectMeta{
@@ -404,7 +420,7 @@ func (r *CFAppReconciler) reconcileVCAPSecret(
 	_, err = controllerutil.CreateOrPatch(ctx, r.k8sClient, secret, func() error {
 		secret.StringData = envValue
 
-		return controllerutil.SetOwnerReference(cfApp, secret, r.scheme)
+		return controllerutil.SetControllerReference(cfApp, secret, r.scheme)
 	})
 	if err != nil {
 		log.Info("unable to create or patch Secret", "reason", err)
